Misc clean ups
