--- src/sections/Landing.tsx
@@ -1,13 +1,0 @@
-import type { FC } from 'react'
-
-import SecondaryNavBar from '../components/SecondaryNavBar'
-
-const Landing: FC = () => {
-  return (
-    <>
-      <SecondaryNavBar />
-    </>
-  )
-}
-
-export default Landing
