+++ src/sections/SampleSection.tsx
@@ -8,11 +8,9 @@ interface InputProps {
 
 const SampleSection: FC<InputProps> = ({ text }) => {
   return (
-    <>
-      <div className="h-[400px] border-spacing-1 text-6xl">
-        <HighlightedTitle title={text} />
-      </div>
-    </>
+    <div className="h-[400px] border-spacing-1 text-6xl">
+      <HighlightedTitle title={text} />
+    </div>
   )
 }
